backend/app/requirements.txt
# SPDX-FileCopyrightText: 2020 - 2024
# - Kotyba Alhaj Taha <kotyba.alhaj-taha@ufz.de>
# - Nils Brinckmann <nils.brinckmann@gfz-potsdam.de>
# - Dirk Ecker <d.ecker@fz-juelich.de>
# - Forschungszentrum Jülich GmbH (FZJ, https://fz-juelich.de)
# - Helmholtz Centre Potsdam - GFZ German Research Centre for Geosciences (GFZ, https://www.gfz-potsdam.de)
# - Helmholtz Centre for Environmental Research GmbH - UFZ (UFZ, https://www.ufz.de)
#
# SPDX-License-Identifier: EUPL-1.2
alembic==1.13.1
cachetools==5.3.2
certifi==2023.11.17
chardet==5.2.0
cherrypicker==0.4.0
click==8.1.7
coverage==7.4.0
cryptography==41.0.7
elasticsearch==8.11.1
entrypoints==0.4
environs==10.2.0
Faker==22.2.0
Flask-Cors==4.0.0
Flask-Executor==1.0.0
Flask-Migrate==4.0.5
Flask-REST-JSONAPI==0.31.2
Flask-SQLAlchemy==2.5.1  # fixed
Flask-Testing==0.8.1
Flask==2.1.3  # fixed
flask_jwt_extended==4.6.0
future==0.18.3
GeoAlchemy2==0.14.3
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow-jsonapi==0.24.0
marshmallow==3.20.2
mccabe==0.7.0
minio==7.2.3
pandas==2.1.4
psycopg2-binary==2.9.9
py-healthcheck==1.10.1
pycodestyle==2.11.1
pycparser==2.21
pycryptodomex==3.20.0
python-dateutil==2.8.2
python-editor==1.0.4
requests==2.31.0
shapely==2.0.2
six==1.16.0
SQLAlchemy-Utils==0.41.1
SQLAlchemy==1.4.50  # fixed
stringcase==1.2.0
tox==4.11.4
Werkzeug==2.3.7  # fixed
xmlschema==1.1.1  # fixed
owslib==0.29.3
